requirements.txt
streamlit>=1.45.0
pandas>=2.0
scikit-learn>=1.4
plotly>=6.0
numpy>=2.0
sentence-transformers
hdbscan>=0.8.33
umap-learn>=0.5.0